Update Purview Catalog SDK for js
