+++ sdk/purview/purview-catalog-rest/rollup.config.js
@@ -1,0 +1,2 @@
+import { makeConfig } from "@azure/dev-tool/shared-config/rollup";
+export default makeConfig(require("./package.json"));
