+++ sdk/purview/purview-catalog-rest/src/index.ts
@@ -6,6 +6,7 @@ import PurviewCatalog from "./purviewCatalog";
 export * from "./purviewCatalog";
 export * from "./parameters";
 export * from "./responses";
+export * from "./clientDefinitions";
 export * from "./models";
 export * from "./outputModels";
 export * from "./pollingHelper";
