+++ sdk/purview/purview-catalog-rest/src/models.ts
@@ -1,18 +1,9 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT license.
 
-export interface AtlasEntityExtInfo {
-  /** The referred entities. */
-  referredEntities?: Record<string, AtlasEntity>;
-}
-
-export interface AtlasStruct {
-  /** The attributes of the struct. */
-  attributes?: Record<string, Record<string, unknown>>;
-  /** The name of the type. */
-  typeName?: string;
-  /** ETag for concurrency control. */
-  lastModifiedTS?: string;
+export interface AtlasEntityWithExtInfo extends AtlasEntityExtInfo {
+  /** An instance of an entity - like hive_table, hive_database. */
+  entity?: AtlasEntity;
 }
 
 export interface AtlasEntity extends AtlasStruct {
@@ -74,6 +65,15 @@ export interface TimeBoundary {
   timeZone?: string;
 }
 
+export interface AtlasStruct {
+  /** The attributes of the struct. */
+  attributes?: Record<string, Record<string, unknown>>;
+  /** The name of the type. */
+  typeName?: string;
+  /** ETag for concurrency control. */
+  lastModifiedTS?: string;
+}
+
 export interface AtlasTermAssignmentHeader {
   /** The confidence of the term assignment. */
   confidence?: number;
@@ -111,9 +111,9 @@ export interface ContactBasic {
   info?: string;
 }
 
-export interface AtlasEntityWithExtInfo extends AtlasEntityExtInfo {
-  /** An instance of an entity - like hive_table, hive_database. */
-  entity?: AtlasEntity;
+export interface AtlasEntityExtInfo {
+  /** The referred entities. */
+  referredEntities?: Record<string, AtlasEntity>;
 }
 
 export interface AtlasEntityHeader extends AtlasStruct {
@@ -148,26 +148,6 @@ export interface ClassificationAssociateRequest {
 export interface AtlasEntityHeaders {
   /** The description of the guid header map, */
   guidHeaderMap?: Record<string, AtlasEntityHeader>;
-}
-
-export interface AtlasBaseModelObject {
-  /** The GUID of the object. */
-  guid?: string;
-}
-
-export interface AtlasGlossaryBaseObject extends AtlasBaseModelObject {
-  /** An array of classifications. */
-  classifications?: Array<AtlasClassification>;
-  /** The long version description. */
-  longDescription?: string;
-  /** The name of the glossary object. */
-  name?: string;
-  /** The qualified name of the glossary object. */
-  qualifiedName?: string;
-  /** The short version of description. */
-  shortDescription?: string;
-  /** ETag for concurrency control. */
-  lastModifiedTS?: string;
 }
 
 export interface AtlasGlossary extends AtlasGlossaryBaseObject {
@@ -211,6 +191,26 @@ export interface AtlasRelatedTermHeader {
   steward?: string;
   /** The GUID of the term. */
   termGuid?: string;
+}
+
+export interface AtlasGlossaryBaseObject extends AtlasBaseModelObject {
+  /** An array of classifications. */
+  classifications?: Array<AtlasClassification>;
+  /** The long version description. */
+  longDescription?: string;
+  /** The name of the glossary object. */
+  name?: string;
+  /** The qualified name of the glossary object. */
+  qualifiedName?: string;
+  /** The short version of description. */
+  shortDescription?: string;
+  /** ETag for concurrency control. */
+  lastModifiedTS?: string;
+}
+
+export interface AtlasBaseModelObject {
+  /** The GUID of the object. */
+  guid?: string;
 }
 
 export interface AtlasGlossaryCategory extends AtlasGlossaryBaseObject {
@@ -301,15 +301,6 @@ export interface ResourceLink {
   url?: string;
 }
 
-export interface AtlasObjectId {
-  /** The GUID of the object. */
-  guid?: string;
-  /** The name of the type. */
-  typeName?: string;
-  /** The unique attributes of the object. */
-  uniqueAttributes?: Record<string, Record<string, unknown>>;
-}
-
 export interface AtlasRelatedObjectId extends AtlasObjectId {
   /** The display text. */
   displayText?: string;
@@ -322,6 +313,15 @@ export interface AtlasRelatedObjectId extends AtlasObjectId {
   relationshipGuid?: string;
   /** The enum of relationship status. */
   relationshipStatus?: "ACTIVE" | "DELETED";
+}
+
+export interface AtlasObjectId {
+  /** The GUID of the object. */
+  guid?: string;
+  /** The name of the type. */
+  typeName?: string;
+  /** The unique attributes of the object. */
+  uniqueAttributes?: Record<string, Record<string, unknown>>;
 }
 
 export interface AtlasTermCategorizationHeader {
@@ -347,7 +347,7 @@ export interface AtlasGlossaryExtInfo extends AtlasGlossary {
 export interface SearchRequest {
   /** The keywords applied to all searchable fields. */
   keywords?: string;
-  /** The offset. The default value is 0. */
+  /** The offset. The default value is 0. The maximum value is 100000. */
   offset?: number;
   /** The limit of the number of the search result. default value is 50; maximum value is 1000. */
   limit?: number;
@@ -386,9 +386,9 @@ export interface BrowseRequest {
   entityType?: string;
   /** The path to browse the next level child entities. */
   path?: string;
-  /** The number of browse items we hope to return. */
+  /** The number of browse items we hope to return. The maximum value is 10000. */
   limit?: number;
-  /** The offset. The default value is 0. */
+  /** The offset. The default value is 0. The maximum value is 100000. */
   offset?: number;
 }
 
@@ -426,6 +426,67 @@ export interface AtlasRelationship extends AtlasStruct {
   updatedBy?: string;
   /** The version of the relationship. */
   version?: number;
+}
+
+export interface AtlasClassificationDef extends AtlasStructDef {
+  /**
+   * Specifying a list of entityType names in the classificationDef, ensures that classifications can
+   * only be applied to those entityTypes.
+   * <ul>
+   * <li>Any subtypes of the entity types inherit the restriction</li>
+   * <li>Any classificationDef subtypes inherit the parents entityTypes restrictions</li>
+   * <li>Any classificationDef subtypes can further restrict the parents entityTypes restrictions by specifying a subset of the entityTypes</li>
+   * <li>An empty entityTypes list when there are no parent restrictions means there are no restrictions</li>
+   * <li>An empty entityTypes list when there are parent restrictions means that the subtype picks up the parents restrictions</li>
+   * <li>If a list of entityTypes are supplied, where one inherits from another, this will be rejected. This should encourage cleaner classificationsDefs</li>
+   * </ul>
+   */
+  entityTypes?: Array<string>;
+  /** An array of sub types. */
+  subTypes?: Array<string>;
+  /** An array of super types. */
+  superTypes?: Array<string>;
+}
+
+export interface AtlasStructDef extends AtlasBaseTypeDef {
+  /** An array of attribute definitions. */
+  attributeDefs?: Array<AtlasAttributeDef>;
+}
+
+export interface AtlasAttributeDef {
+  /** single-valued attribute or multi-valued attribute. */
+  cardinality?: "SINGLE" | "LIST" | "SET";
+  /** An array of constraints. */
+  constraints?: Array<AtlasConstraintDef>;
+  /** The default value of the attribute. */
+  defaultValue?: string;
+  /** The description of the attribute. */
+  description?: string;
+  /** Determines if it is included in notification. */
+  includeInNotification?: boolean;
+  /** Determines if it is indexable. */
+  isIndexable?: boolean;
+  /** Determines if it is optional. */
+  isOptional?: boolean;
+  /** Determines if it unique. */
+  isUnique?: boolean;
+  /** The name of the attribute. */
+  name?: string;
+  /** The options for the attribute. */
+  options?: Record<string, string>;
+  /** The name of the type. */
+  typeName?: string;
+  /** The maximum count of the values. */
+  valuesMaxCount?: number;
+  /** The minimum count of the values. */
+  valuesMinCount?: number;
+}
+
+export interface AtlasConstraintDef {
+  /** The parameters of the constraint definition. */
+  params?: Record<string, Record<string, unknown>>;
+  /** The type of the constraint. */
+  type?: string;
 }
 
 export interface AtlasBaseTypeDef {
@@ -543,67 +604,6 @@ export interface TimeZone {
   rawOffset?: number;
 }
 
-export interface AtlasStructDef extends AtlasBaseTypeDef {
-  /** An array of attribute definitions. */
-  attributeDefs?: Array<AtlasAttributeDef>;
-}
-
-export interface AtlasAttributeDef {
-  /** single-valued attribute or multi-valued attribute. */
-  cardinality?: "SINGLE" | "LIST" | "SET";
-  /** An array of constraints. */
-  constraints?: Array<AtlasConstraintDef>;
-  /** The default value of the attribute. */
-  defaultValue?: string;
-  /** The description of the attribute. */
-  description?: string;
-  /** Determines if it is included in notification. */
-  includeInNotification?: boolean;
-  /** Determines if it is indexable. */
-  isIndexable?: boolean;
-  /** Determines if it is optional. */
-  isOptional?: boolean;
-  /** Determines if it unique. */
-  isUnique?: boolean;
-  /** The name of the attribute. */
-  name?: string;
-  /** The options for the attribute. */
-  options?: Record<string, string>;
-  /** The name of the type. */
-  typeName?: string;
-  /** The maximum count of the values. */
-  valuesMaxCount?: number;
-  /** The minimum count of the values. */
-  valuesMinCount?: number;
-}
-
-export interface AtlasConstraintDef {
-  /** The parameters of the constraint definition. */
-  params?: Record<string, Record<string, unknown>>;
-  /** The type of the constraint. */
-  type?: string;
-}
-
-export interface AtlasClassificationDef extends AtlasStructDef {
-  /**
-   * Specifying a list of entityType names in the classificationDef, ensures that classifications can
-   * only be applied to those entityTypes.
-   * <ul>
-   * <li>Any subtypes of the entity types inherit the restriction</li>
-   * <li>Any classificationDef subtypes inherit the parents entityTypes restrictions</li>
-   * <li>Any classificationDef subtypes can further restrict the parents entityTypes restrictions by specifying a subset of the entityTypes</li>
-   * <li>An empty entityTypes list when there are no parent restrictions means there are no restrictions</li>
-   * <li>An empty entityTypes list when there are parent restrictions means that the subtype picks up the parents restrictions</li>
-   * <li>If a list of entityTypes are supplied, where one inherits from another, this will be rejected. This should encourage cleaner classificationsDefs</li>
-   * </ul>
-   */
-  entityTypes?: Array<string>;
-  /** An array of sub types. */
-  subTypes?: Array<string>;
-  /** An array of super types. */
-  superTypes?: Array<string>;
-}
-
 export interface AtlasEntityDef extends AtlasStructDef {
   /** An array of sub types. */
   subTypes?: Array<string>;
@@ -678,6 +678,8 @@ export interface AtlasRelationshipEndDef {
   type?: string;
 }
 
+export interface AtlasTypeDef extends AtlasBaseTypeDef, AtlasExtraTypeDef {}
+
 export interface AtlasExtraTypeDef {
   /**
    * Specifying a list of entityType names in the classificationDef, ensures that classifications can
@@ -729,8 +731,6 @@ export interface AtlasExtraTypeDef {
   /** An array of attribute definitions. */
   attributeDefs?: Array<AtlasAttributeDef>;
 }
-
-export interface AtlasTypeDef extends AtlasBaseTypeDef, AtlasExtraTypeDef {}
 
 export interface AtlasTypesDef {
   /** An array of classification definitions. */
