+++ sdk/purview/purview-catalog-rest/src/outputModels.ts
@@ -1,18 +1,9 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT license.
 
-export interface AtlasEntityExtInfoOutput {
-  /** The referred entities. */
-  referredEntities?: Record<string, AtlasEntityOutput>;
-}
-
-export interface AtlasStructOutput {
-  /** The attributes of the struct. */
-  attributes?: Record<string, Record<string, unknown>>;
-  /** The name of the type. */
-  typeName?: string;
-  /** ETag for concurrency control. */
-  lastModifiedTS?: string;
+export interface AtlasEntityWithExtInfoOutput extends AtlasEntityExtInfoOutput {
+  /** An instance of an entity - like hive_table, hive_database. */
+  entity?: AtlasEntityOutput;
 }
 
 export interface AtlasEntityOutput extends AtlasStructOutput {
@@ -74,6 +65,15 @@ export interface TimeBoundaryOutput {
   timeZone?: string;
 }
 
+export interface AtlasStructOutput {
+  /** The attributes of the struct. */
+  attributes?: Record<string, Record<string, unknown>>;
+  /** The name of the type. */
+  typeName?: string;
+  /** ETag for concurrency control. */
+  lastModifiedTS?: string;
+}
+
 export interface AtlasTermAssignmentHeaderOutput {
   /** The confidence of the term assignment. */
   confidence?: number;
@@ -111,9 +111,9 @@ export interface ContactBasicOutput {
   info?: string;
 }
 
-export interface AtlasEntityWithExtInfoOutput extends AtlasEntityExtInfoOutput {
-  /** An instance of an entity - like hive_table, hive_database. */
-  entity?: AtlasEntityOutput;
+export interface AtlasEntityExtInfoOutput {
+  /** The referred entities. */
+  referredEntities?: Record<string, AtlasEntityOutput>;
 }
 
 export interface EntityMutationResponseOutput {
@@ -151,10 +151,13 @@ export interface ErrorResponseOutput {
   errorMessage?: string;
 }
 
-export interface AtlasEntitiesWithExtInfoOutput extends AtlasEntityExtInfoOutput {
+export interface AtlasEntitiesWithExtInfoOutput
+  extends AtlasEntityExtInfoOutput {
   /** An array of entities. */
   entities?: Array<AtlasEntityOutput>;
 }
+
+export interface AtlasClassificationsOutput extends PListOutput {}
 
 export interface PListOutput {
   /** An array of objects. */
@@ -169,28 +172,6 @@ export interface PListOutput {
   startIndex?: number;
   /** The total count of items. */
   totalCount?: number;
-}
-
-export interface AtlasClassificationsOutput extends PListOutput {}
-
-export interface AtlasBaseModelObjectOutput {
-  /** The GUID of the object. */
-  guid?: string;
-}
-
-export interface AtlasGlossaryBaseObjectOutput extends AtlasBaseModelObjectOutput {
-  /** An array of classifications. */
-  classifications?: Array<AtlasClassificationOutput>;
-  /** The long version description. */
-  longDescription?: string;
-  /** The name of the glossary object. */
-  name?: string;
-  /** The qualified name of the glossary object. */
-  qualifiedName?: string;
-  /** The short version of description. */
-  shortDescription?: string;
-  /** ETag for concurrency control. */
-  lastModifiedTS?: string;
 }
 
 export interface AtlasGlossaryOutput extends AtlasGlossaryBaseObjectOutput {
@@ -236,7 +217,29 @@ export interface AtlasRelatedTermHeaderOutput {
   termGuid?: string;
 }
 
-export interface AtlasGlossaryCategoryOutput extends AtlasGlossaryBaseObjectOutput {
+export interface AtlasGlossaryBaseObjectOutput
+  extends AtlasBaseModelObjectOutput {
+  /** An array of classifications. */
+  classifications?: Array<AtlasClassificationOutput>;
+  /** The long version description. */
+  longDescription?: string;
+  /** The name of the glossary object. */
+  name?: string;
+  /** The qualified name of the glossary object. */
+  qualifiedName?: string;
+  /** The short version of description. */
+  shortDescription?: string;
+  /** ETag for concurrency control. */
+  lastModifiedTS?: string;
+}
+
+export interface AtlasBaseModelObjectOutput {
+  /** The GUID of the object. */
+  guid?: string;
+}
+
+export interface AtlasGlossaryCategoryOutput
+  extends AtlasGlossaryBaseObjectOutput {
   /** The glossary header with basic information. */
   anchor?: AtlasGlossaryHeaderOutput;
   /** An array of children categories. */
@@ -324,15 +327,6 @@ export interface ResourceLinkOutput {
   url?: string;
 }
 
-export interface AtlasObjectIdOutput {
-  /** The GUID of the object. */
-  guid?: string;
-  /** The name of the type. */
-  typeName?: string;
-  /** The unique attributes of the object. */
-  uniqueAttributes?: Record<string, Record<string, unknown>>;
-}
-
 export interface AtlasRelatedObjectIdOutput extends AtlasObjectIdOutput {
   /** The display text. */
   displayText?: string;
@@ -345,6 +339,15 @@ export interface AtlasRelatedObjectIdOutput extends AtlasObjectIdOutput {
   relationshipGuid?: string;
   /** The enum of relationship status. */
   relationshipStatus?: "ACTIVE" | "DELETED";
+}
+
+export interface AtlasObjectIdOutput {
+  /** The GUID of the object. */
+  guid?: string;
+  /** The name of the type. */
+  typeName?: string;
+  /** The unique attributes of the object. */
+  uniqueAttributes?: Record<string, Record<string, unknown>>;
 }
 
 export interface AtlasTermCategorizationHeaderOutput {
@@ -397,7 +400,7 @@ export interface ImportCSVOperationErrorOutput {
 export interface SearchResultOutput {
   /** The total number of search results (not the number of documents in a single page). */
   "@search.count"?: number;
-  /** A facet list that consists of index fields assetType ,classification, classificationCategory, contactId, fileExtension, label, and label. When the facet is specified in the request, the value of the facet is returned as an element of @search.facets. */
+  /** A facet list that consists of index fields assetType ,classification, contactId, and label. When the facet is specified in the request, the value of the facet is returned as an element of @search.facets. */
   "@search.facets"?: SearchFacetResultValueOutput;
   value?: Array<SearchResultValueOutput>;
 }
@@ -629,6 +632,67 @@ export interface AtlasRelationshipWithExtInfoOutput {
   relationship?: AtlasRelationshipOutput;
 }
 
+export interface AtlasClassificationDefOutput extends AtlasStructDefOutput {
+  /**
+   * Specifying a list of entityType names in the classificationDef, ensures that classifications can
+   * only be applied to those entityTypes.
+   * <ul>
+   * <li>Any subtypes of the entity types inherit the restriction</li>
+   * <li>Any classificationDef subtypes inherit the parents entityTypes restrictions</li>
+   * <li>Any classificationDef subtypes can further restrict the parents entityTypes restrictions by specifying a subset of the entityTypes</li>
+   * <li>An empty entityTypes list when there are no parent restrictions means there are no restrictions</li>
+   * <li>An empty entityTypes list when there are parent restrictions means that the subtype picks up the parents restrictions</li>
+   * <li>If a list of entityTypes are supplied, where one inherits from another, this will be rejected. This should encourage cleaner classificationsDefs</li>
+   * </ul>
+   */
+  entityTypes?: Array<string>;
+  /** An array of sub types. */
+  subTypes?: Array<string>;
+  /** An array of super types. */
+  superTypes?: Array<string>;
+}
+
+export interface AtlasStructDefOutput extends AtlasBaseTypeDefOutput {
+  /** An array of attribute definitions. */
+  attributeDefs?: Array<AtlasAttributeDefOutput>;
+}
+
+export interface AtlasAttributeDefOutput {
+  /** single-valued attribute or multi-valued attribute. */
+  cardinality?: "SINGLE" | "LIST" | "SET";
+  /** An array of constraints. */
+  constraints?: Array<AtlasConstraintDefOutput>;
+  /** The default value of the attribute. */
+  defaultValue?: string;
+  /** The description of the attribute. */
+  description?: string;
+  /** Determines if it is included in notification. */
+  includeInNotification?: boolean;
+  /** Determines if it is indexable. */
+  isIndexable?: boolean;
+  /** Determines if it is optional. */
+  isOptional?: boolean;
+  /** Determines if it unique. */
+  isUnique?: boolean;
+  /** The name of the attribute. */
+  name?: string;
+  /** The options for the attribute. */
+  options?: Record<string, string>;
+  /** The name of the type. */
+  typeName?: string;
+  /** The maximum count of the values. */
+  valuesMaxCount?: number;
+  /** The minimum count of the values. */
+  valuesMinCount?: number;
+}
+
+export interface AtlasConstraintDefOutput {
+  /** The parameters of the constraint definition. */
+  params?: Record<string, Record<string, unknown>>;
+  /** The type of the constraint. */
+  type?: string;
+}
+
 export interface AtlasBaseTypeDefOutput {
   /** The enum of type category. */
   category?:
@@ -744,67 +808,6 @@ export interface TimeZoneOutput {
   rawOffset?: number;
 }
 
-export interface AtlasStructDefOutput extends AtlasBaseTypeDefOutput {
-  /** An array of attribute definitions. */
-  attributeDefs?: Array<AtlasAttributeDefOutput>;
-}
-
-export interface AtlasAttributeDefOutput {
-  /** single-valued attribute or multi-valued attribute. */
-  cardinality?: "SINGLE" | "LIST" | "SET";
-  /** An array of constraints. */
-  constraints?: Array<AtlasConstraintDefOutput>;
-  /** The default value of the attribute. */
-  defaultValue?: string;
-  /** The description of the attribute. */
-  description?: string;
-  /** Determines if it is included in notification. */
-  includeInNotification?: boolean;
-  /** Determines if it is indexable. */
-  isIndexable?: boolean;
-  /** Determines if it is optional. */
-  isOptional?: boolean;
-  /** Determines if it unique. */
-  isUnique?: boolean;
-  /** The name of the attribute. */
-  name?: string;
-  /** The options for the attribute. */
-  options?: Record<string, string>;
-  /** The name of the type. */
-  typeName?: string;
-  /** The maximum count of the values. */
-  valuesMaxCount?: number;
-  /** The minimum count of the values. */
-  valuesMinCount?: number;
-}
-
-export interface AtlasConstraintDefOutput {
-  /** The parameters of the constraint definition. */
-  params?: Record<string, Record<string, unknown>>;
-  /** The type of the constraint. */
-  type?: string;
-}
-
-export interface AtlasClassificationDefOutput extends AtlasStructDefOutput {
-  /**
-   * Specifying a list of entityType names in the classificationDef, ensures that classifications can
-   * only be applied to those entityTypes.
-   * <ul>
-   * <li>Any subtypes of the entity types inherit the restriction</li>
-   * <li>Any classificationDef subtypes inherit the parents entityTypes restrictions</li>
-   * <li>Any classificationDef subtypes can further restrict the parents entityTypes restrictions by specifying a subset of the entityTypes</li>
-   * <li>An empty entityTypes list when there are no parent restrictions means there are no restrictions</li>
-   * <li>An empty entityTypes list when there are parent restrictions means that the subtype picks up the parents restrictions</li>
-   * <li>If a list of entityTypes are supplied, where one inherits from another, this will be rejected. This should encourage cleaner classificationsDefs</li>
-   * </ul>
-   */
-  entityTypes?: Array<string>;
-  /** An array of sub types. */
-  subTypes?: Array<string>;
-  /** An array of super types. */
-  superTypes?: Array<string>;
-}
-
 export interface AtlasEntityDefOutput extends AtlasStructDefOutput {
   /** An array of sub types. */
   subTypes?: Array<string>;
@@ -814,7 +817,8 @@ export interface AtlasEntityDefOutput extends AtlasStructDefOutput {
   relationshipAttributeDefs?: Array<AtlasRelationshipAttributeDefOutput>;
 }
 
-export interface AtlasRelationshipAttributeDefOutput extends AtlasAttributeDefOutput {
+export interface AtlasRelationshipAttributeDefOutput
+  extends AtlasAttributeDefOutput {
   /** Determines if it is a legacy attribute. */
   isLegacyAttribute?: boolean;
   /** The name of the relationship type. */
@@ -879,6 +883,10 @@ export interface AtlasRelationshipEndDefOutput {
   type?: string;
 }
 
+export interface AtlasTypeDefOutput
+  extends AtlasBaseTypeDefOutput,
+    AtlasExtraTypeDefOutput {}
+
 export interface AtlasExtraTypeDefOutput {
   /**
    * Specifying a list of entityType names in the classificationDef, ensures that classifications can
@@ -930,8 +938,6 @@ export interface AtlasExtraTypeDefOutput {
   /** An array of attribute definitions. */
   attributeDefs?: Array<AtlasAttributeDefOutput>;
 }
-
-export interface AtlasTypeDefOutput extends AtlasBaseTypeDefOutput, AtlasExtraTypeDefOutput {}
 
 export interface AtlasTypesDefOutput {
   /** An array of classification definitions. */
