+++ sdk/purview/purview-catalog-rest/src/parameters.ts
@@ -18,7 +18,7 @@ import {
   AutoCompleteRequest,
   AtlasRelationship,
   AtlasTypesDef,
-  MoveEntitiesRequest,
+  MoveEntitiesRequest
 } from "./models";
 
 export interface EntityCreateOrUpdateBodyParam {
@@ -26,11 +26,18 @@ export interface EntityCreateOrUpdateBodyParam {
   body: AtlasEntityWithExtInfo;
 }
 
-export type EntityCreateOrUpdateParameters = EntityCreateOrUpdateBodyParam & RequestParameters;
+export interface EntityCreateOrUpdateMediaTypesParam {
+  /** Request content type */
+  contentType?: "application/json";
+}
+
+export type EntityCreateOrUpdateParameters = EntityCreateOrUpdateMediaTypesParam &
+  EntityCreateOrUpdateBodyParam &
+  RequestParameters;
 
 export interface EntityListByGuidsQueryParamProperties {
-  /** An array of GUIDs of entities to create. */
-  guids: Array<string>;
+  /** An array of GUIDs of entities to list. */
+  guid: Array<string>;
   /** Whether to return minimal information for referred entities. */
   minExtInfo?: boolean;
   /** Whether to ignore relationship attributes. */
@@ -43,33 +50,47 @@ export interface EntityListByGuidsQueryParam {
   queryParameters: EntityListByGuidsQueryParamProperties;
 }
 
-export type EntityListByGuidsParameters = EntityListByGuidsQueryParam & RequestParameters;
+export type EntityListByGuidsParameters = EntityListByGuidsQueryParam &
+  RequestParameters;
 
 export interface EntityCreateOrUpdateEntitiesBodyParam {
   /** An array of entities to create or update. */
   body: AtlasEntitiesWithExtInfo;
 }
 
-export type EntityCreateOrUpdateEntitiesParameters = EntityCreateOrUpdateEntitiesBodyParam &
+export interface EntityCreateOrUpdateEntitiesMediaTypesParam {
+  /** Request content type */
+  contentType?: "application/json";
+}
+
+export type EntityCreateOrUpdateEntitiesParameters = EntityCreateOrUpdateEntitiesMediaTypesParam &
+  EntityCreateOrUpdateEntitiesBodyParam &
   RequestParameters;
 
 export interface EntityDeleteByGuidsQueryParamProperties {
   /** An array of GUIDs of entities to delete. */
-  guids: Array<string>;
+  guid: Array<string>;
 }
 
 export interface EntityDeleteByGuidsQueryParam {
   queryParameters: EntityDeleteByGuidsQueryParamProperties;
 }
 
-export type EntityDeleteByGuidsParameters = EntityDeleteByGuidsQueryParam & RequestParameters;
+export type EntityDeleteByGuidsParameters = EntityDeleteByGuidsQueryParam &
+  RequestParameters;
 
 export interface EntityAddClassificationBodyParam {
   /** The request to associate a classification to multiple entities. */
   body: ClassificationAssociateRequest;
 }
 
-export type EntityAddClassificationParameters = EntityAddClassificationBodyParam &
+export interface EntityAddClassificationMediaTypesParam {
+  /** Request content type */
+  contentType?: "application/json";
+}
+
+export type EntityAddClassificationParameters = EntityAddClassificationMediaTypesParam &
+  EntityAddClassificationBodyParam &
   RequestParameters;
 
 export interface EntityGetByGuidQueryParamProperties {
@@ -83,7 +104,8 @@ export interface EntityGetByGuidQueryParam {
   queryParameters?: EntityGetByGuidQueryParamProperties;
 }
 
-export type EntityGetByGuidParameters = EntityGetByGuidQueryParam & RequestParameters;
+export type EntityGetByGuidParameters = EntityGetByGuidQueryParam &
+  RequestParameters;
 
 export interface EntityPartialUpdateEntityAttributeByGuidBodyParam {
   /** The value of the attribute. */
@@ -99,10 +121,15 @@ export interface EntityPartialUpdateEntityAttributeByGuidQueryParam {
   queryParameters: EntityPartialUpdateEntityAttributeByGuidQueryParamProperties;
 }
 
-export type EntityPartialUpdateEntityAttributeByGuidParameters =
-  EntityPartialUpdateEntityAttributeByGuidQueryParam &
-    EntityPartialUpdateEntityAttributeByGuidBodyParam &
-    RequestParameters;
+export interface EntityPartialUpdateEntityAttributeByGuidMediaTypesParam {
+  /** Request content type */
+  contentType?: "application/json";
+}
+
+export type EntityPartialUpdateEntityAttributeByGuidParameters = EntityPartialUpdateEntityAttributeByGuidQueryParam &
+  EntityPartialUpdateEntityAttributeByGuidMediaTypesParam &
+  EntityPartialUpdateEntityAttributeByGuidBodyParam &
+  RequestParameters;
 export type EntityDeleteByGuidParameters = RequestParameters;
 export type EntityGetClassificationParameters = RequestParameters;
 export type EntityDeleteClassificationParameters = RequestParameters;
@@ -113,7 +140,13 @@ export interface EntityAddClassificationsBodyParam {
   body: Array<AtlasClassification>;
 }
 
-export type EntityAddClassificationsParameters = EntityAddClassificationsBodyParam &
+export interface EntityAddClassificationsMediaTypesParam {
+  /** Request content type */
+  contentType?: "application/json";
+}
+
+export type EntityAddClassificationsParameters = EntityAddClassificationsMediaTypesParam &
+  EntityAddClassificationsBodyParam &
   RequestParameters;
 
 export interface EntityUpdateClassificationsBodyParam {
@@ -121,7 +154,13 @@ export interface EntityUpdateClassificationsBodyParam {
   body: Array<AtlasClassification>;
 }
 
-export type EntityUpdateClassificationsParameters = EntityUpdateClassificationsBodyParam &
+export interface EntityUpdateClassificationsMediaTypesParam {
+  /** Request content type */
+  contentType?: "application/json";
+}
+
+export type EntityUpdateClassificationsParameters = EntityUpdateClassificationsMediaTypesParam &
+  EntityUpdateClassificationsBodyParam &
   RequestParameters;
 
 export interface EntityGetByUniqueAttributesQueryParamProperties {
@@ -154,10 +193,15 @@ export interface EntityPartialUpdateEntityByUniqueAttributesQueryParam {
   queryParameters?: EntityPartialUpdateEntityByUniqueAttributesQueryParamProperties;
 }
 
-export type EntityPartialUpdateEntityByUniqueAttributesParameters =
-  EntityPartialUpdateEntityByUniqueAttributesQueryParam &
-    EntityPartialUpdateEntityByUniqueAttributesBodyParam &
-    RequestParameters;
+export interface EntityPartialUpdateEntityByUniqueAttributesMediaTypesParam {
+  /** Request content type */
+  contentType?: "application/json";
+}
+
+export type EntityPartialUpdateEntityByUniqueAttributesParameters = EntityPartialUpdateEntityByUniqueAttributesQueryParam &
+  EntityPartialUpdateEntityByUniqueAttributesMediaTypesParam &
+  EntityPartialUpdateEntityByUniqueAttributesBodyParam &
+  RequestParameters;
 
 export interface EntityDeleteByUniqueAttributeQueryParamProperties {
   /** The qualified name of the entity. */
@@ -180,8 +224,8 @@ export interface EntityDeleteClassificationByUniqueAttributeQueryParam {
   queryParameters?: EntityDeleteClassificationByUniqueAttributeQueryParamProperties;
 }
 
-export type EntityDeleteClassificationByUniqueAttributeParameters =
-  EntityDeleteClassificationByUniqueAttributeQueryParam & RequestParameters;
+export type EntityDeleteClassificationByUniqueAttributeParameters = EntityDeleteClassificationByUniqueAttributeQueryParam &
+  RequestParameters;
 
 export interface EntityAddClassificationsByUniqueAttributeBodyParam {
   /** An array of classification to be added. */
@@ -197,10 +241,15 @@ export interface EntityAddClassificationsByUniqueAttributeQueryParam {
   queryParameters?: EntityAddClassificationsByUniqueAttributeQueryParamProperties;
 }
 
-export type EntityAddClassificationsByUniqueAttributeParameters =
-  EntityAddClassificationsByUniqueAttributeQueryParam &
-    EntityAddClassificationsByUniqueAttributeBodyParam &
-    RequestParameters;
+export interface EntityAddClassificationsByUniqueAttributeMediaTypesParam {
+  /** Request content type */
+  contentType?: "application/json";
+}
+
+export type EntityAddClassificationsByUniqueAttributeParameters = EntityAddClassificationsByUniqueAttributeQueryParam &
+  EntityAddClassificationsByUniqueAttributeMediaTypesParam &
+  EntityAddClassificationsByUniqueAttributeBodyParam &
+  RequestParameters;
 
 export interface EntityUpdateClassificationsByUniqueAttributeBodyParam {
   /** An array of classification to be updated. */
@@ -216,17 +265,28 @@ export interface EntityUpdateClassificationsByUniqueAttributeQueryParam {
   queryParameters?: EntityUpdateClassificationsByUniqueAttributeQueryParamProperties;
 }
 
-export type EntityUpdateClassificationsByUniqueAttributeParameters =
-  EntityUpdateClassificationsByUniqueAttributeQueryParam &
-    EntityUpdateClassificationsByUniqueAttributeBodyParam &
-    RequestParameters;
+export interface EntityUpdateClassificationsByUniqueAttributeMediaTypesParam {
+  /** Request content type */
+  contentType?: "application/json";
+}
+
+export type EntityUpdateClassificationsByUniqueAttributeParameters = EntityUpdateClassificationsByUniqueAttributeQueryParam &
+  EntityUpdateClassificationsByUniqueAttributeMediaTypesParam &
+  EntityUpdateClassificationsByUniqueAttributeBodyParam &
+  RequestParameters;
 
 export interface EntitySetClassificationsBodyParam {
   /** Atlas entity headers. */
   body: AtlasEntityHeaders;
 }
 
-export type EntitySetClassificationsParameters = EntitySetClassificationsBodyParam &
+export interface EntitySetClassificationsMediaTypesParam {
+  /** Request content type */
+  contentType?: "application/json";
+}
+
+export type EntitySetClassificationsParameters = EntitySetClassificationsMediaTypesParam &
+  EntitySetClassificationsBodyParam &
   RequestParameters;
 
 export interface EntityGetEntitiesByUniqueAttributesQueryParamProperties {
@@ -242,8 +302,8 @@ export interface EntityGetEntitiesByUniqueAttributesQueryParam {
   queryParameters?: EntityGetEntitiesByUniqueAttributesQueryParamProperties;
 }
 
-export type EntityGetEntitiesByUniqueAttributesParameters =
-  EntityGetEntitiesByUniqueAttributesQueryParam & RequestParameters;
+export type EntityGetEntitiesByUniqueAttributesParameters = EntityGetEntitiesByUniqueAttributesQueryParam &
+  RequestParameters;
 export type EntityGetHeaderParameters = RequestParameters;
 
 export interface GlossaryListGlossariesQueryParamProperties {
@@ -261,7 +321,8 @@ export interface GlossaryListGlossariesQueryParam {
   queryParameters?: GlossaryListGlossariesQueryParamProperties;
 }
 
-export type GlossaryListGlossariesParameters = GlossaryListGlossariesQueryParam & RequestParameters;
+export type GlossaryListGlossariesParameters = GlossaryListGlossariesQueryParam &
+  RequestParameters;
 
 export interface GlossaryCreateGlossaryBodyParam {
   /**
@@ -271,14 +332,27 @@ export interface GlossaryCreateGlossaryBodyParam {
   body: AtlasGlossary;
 }
 
-export type GlossaryCreateGlossaryParameters = GlossaryCreateGlossaryBodyParam & RequestParameters;
+export interface GlossaryCreateGlossaryMediaTypesParam {
+  /** Request content type */
+  contentType?: "application/json";
+}
+
+export type GlossaryCreateGlossaryParameters = GlossaryCreateGlossaryMediaTypesParam &
+  GlossaryCreateGlossaryBodyParam &
+  RequestParameters;
 
 export interface GlossaryCreateGlossaryCategoriesBodyParam {
   /** An array of glossary category definitions to be created. */
   body: Array<AtlasGlossaryCategory>;
 }
 
-export type GlossaryCreateGlossaryCategoriesParameters = GlossaryCreateGlossaryCategoriesBodyParam &
+export interface GlossaryCreateGlossaryCategoriesMediaTypesParam {
+  /** Request content type */
+  contentType?: "application/json";
+}
+
+export type GlossaryCreateGlossaryCategoriesParameters = GlossaryCreateGlossaryCategoriesMediaTypesParam &
+  GlossaryCreateGlossaryCategoriesBodyParam &
   RequestParameters;
 
 export interface GlossaryCreateGlossaryCategoryBodyParam {
@@ -289,7 +363,13 @@ export interface GlossaryCreateGlossaryCategoryBodyParam {
   body: AtlasGlossaryCategory;
 }
 
-export type GlossaryCreateGlossaryCategoryParameters = GlossaryCreateGlossaryCategoryBodyParam &
+export interface GlossaryCreateGlossaryCategoryMediaTypesParam {
+  /** Request content type */
+  contentType?: "application/json";
+}
+
+export type GlossaryCreateGlossaryCategoryParameters = GlossaryCreateGlossaryCategoryMediaTypesParam &
+  GlossaryCreateGlossaryCategoryBodyParam &
   RequestParameters;
 export type GlossaryGetGlossaryCategoryParameters = RequestParameters;
 
@@ -298,7 +378,13 @@ export interface GlossaryUpdateGlossaryCategoryBodyParam {
   body: AtlasGlossaryCategory;
 }
 
-export type GlossaryUpdateGlossaryCategoryParameters = GlossaryUpdateGlossaryCategoryBodyParam &
+export interface GlossaryUpdateGlossaryCategoryMediaTypesParam {
+  /** Request content type */
+  contentType?: "application/json";
+}
+
+export type GlossaryUpdateGlossaryCategoryParameters = GlossaryUpdateGlossaryCategoryMediaTypesParam &
+  GlossaryUpdateGlossaryCategoryBodyParam &
   RequestParameters;
 export type GlossaryDeleteGlossaryCategoryParameters = RequestParameters;
 
@@ -307,8 +393,14 @@ export interface GlossaryPartialUpdateGlossaryCategoryBodyParam {
   body: Record<string, string>;
 }
 
-export type GlossaryPartialUpdateGlossaryCategoryParameters =
-  GlossaryPartialUpdateGlossaryCategoryBodyParam & RequestParameters;
+export interface GlossaryPartialUpdateGlossaryCategoryMediaTypesParam {
+  /** Request content type */
+  contentType?: "application/json";
+}
+
+export type GlossaryPartialUpdateGlossaryCategoryParameters = GlossaryPartialUpdateGlossaryCategoryMediaTypesParam &
+  GlossaryPartialUpdateGlossaryCategoryBodyParam &
+  RequestParameters;
 
 export interface GlossaryListRelatedCategoriesQueryParamProperties {
   /** The page size - by default there is no paging. */
@@ -359,7 +451,13 @@ export interface GlossaryCreateGlossaryTermQueryParam {
   queryParameters?: GlossaryCreateGlossaryTermQueryParamProperties;
 }
 
+export interface GlossaryCreateGlossaryTermMediaTypesParam {
+  /** Request content type */
+  contentType?: "application/json";
+}
+
 export type GlossaryCreateGlossaryTermParameters = GlossaryCreateGlossaryTermQueryParam &
+  GlossaryCreateGlossaryTermMediaTypesParam &
   GlossaryCreateGlossaryTermBodyParam &
   RequestParameters;
 
@@ -380,7 +478,13 @@ export interface GlossaryUpdateGlossaryTermBodyParam {
   body: AtlasGlossaryTerm;
 }
 
-export type GlossaryUpdateGlossaryTermParameters = GlossaryUpdateGlossaryTermBodyParam &
+export interface GlossaryUpdateGlossaryTermMediaTypesParam {
+  /** Request content type */
+  contentType?: "application/json";
+}
+
+export type GlossaryUpdateGlossaryTermParameters = GlossaryUpdateGlossaryTermMediaTypesParam &
+  GlossaryUpdateGlossaryTermBodyParam &
   RequestParameters;
 export type GlossaryDeleteGlossaryTermParameters = RequestParameters;
 
@@ -398,10 +502,15 @@ export interface GlossaryPartialUpdateGlossaryTermQueryParam {
   queryParameters?: GlossaryPartialUpdateGlossaryTermQueryParamProperties;
 }
 
-export type GlossaryPartialUpdateGlossaryTermParameters =
-  GlossaryPartialUpdateGlossaryTermQueryParam &
-    GlossaryPartialUpdateGlossaryTermBodyParam &
-    RequestParameters;
+export interface GlossaryPartialUpdateGlossaryTermMediaTypesParam {
+  /** Request content type */
+  contentType?: "application/json";
+}
+
+export type GlossaryPartialUpdateGlossaryTermParameters = GlossaryPartialUpdateGlossaryTermQueryParam &
+  GlossaryPartialUpdateGlossaryTermMediaTypesParam &
+  GlossaryPartialUpdateGlossaryTermBodyParam &
+  RequestParameters;
 
 export interface GlossaryCreateGlossaryTermsBodyParam {
   /** An array of glossary term definitions to be created in bulk. */
@@ -417,7 +526,13 @@ export interface GlossaryCreateGlossaryTermsQueryParam {
   queryParameters?: GlossaryCreateGlossaryTermsQueryParamProperties;
 }
 
+export interface GlossaryCreateGlossaryTermsMediaTypesParam {
+  /** Request content type */
+  contentType?: "application/json";
+}
+
 export type GlossaryCreateGlossaryTermsParameters = GlossaryCreateGlossaryTermsQueryParam &
+  GlossaryCreateGlossaryTermsMediaTypesParam &
   GlossaryCreateGlossaryTermsBodyParam &
   RequestParameters;
 
@@ -434,15 +549,21 @@ export interface GlossaryGetEntitiesAssignedWithTermQueryParam {
   queryParameters?: GlossaryGetEntitiesAssignedWithTermQueryParamProperties;
 }
 
-export type GlossaryGetEntitiesAssignedWithTermParameters =
-  GlossaryGetEntitiesAssignedWithTermQueryParam & RequestParameters;
+export type GlossaryGetEntitiesAssignedWithTermParameters = GlossaryGetEntitiesAssignedWithTermQueryParam &
+  RequestParameters;
 
 export interface GlossaryAssignTermToEntitiesBodyParam {
   /** An array of related object IDs to which the term has to be associated. */
   body: Array<AtlasRelatedObjectId>;
 }
 
-export type GlossaryAssignTermToEntitiesParameters = GlossaryAssignTermToEntitiesBodyParam &
+export interface GlossaryAssignTermToEntitiesMediaTypesParam {
+  /** Request content type */
+  contentType?: "application/json";
+}
+
+export type GlossaryAssignTermToEntitiesParameters = GlossaryAssignTermToEntitiesMediaTypesParam &
+  GlossaryAssignTermToEntitiesBodyParam &
   RequestParameters;
 
 export interface GlossaryRemoveTermAssignmentFromEntitiesBodyParam {
@@ -450,16 +571,28 @@ export interface GlossaryRemoveTermAssignmentFromEntitiesBodyParam {
   body: Array<AtlasRelatedObjectId>;
 }
 
-export type GlossaryRemoveTermAssignmentFromEntitiesParameters =
-  GlossaryRemoveTermAssignmentFromEntitiesBodyParam & RequestParameters;
+export interface GlossaryRemoveTermAssignmentFromEntitiesMediaTypesParam {
+  /** Request content type */
+  contentType?: "application/json";
+}
+
+export type GlossaryRemoveTermAssignmentFromEntitiesParameters = GlossaryRemoveTermAssignmentFromEntitiesMediaTypesParam &
+  GlossaryRemoveTermAssignmentFromEntitiesBodyParam &
+  RequestParameters;
 
 export interface GlossaryDeleteTermAssignmentFromEntitiesBodyParam {
   /** An array of related object IDs from which the term has to be dissociated. */
   body: Array<AtlasRelatedObjectId>;
 }
 
-export type GlossaryDeleteTermAssignmentFromEntitiesParameters =
-  GlossaryDeleteTermAssignmentFromEntitiesBodyParam & RequestParameters;
+export interface GlossaryDeleteTermAssignmentFromEntitiesMediaTypesParam {
+  /** Request content type */
+  contentType?: "application/json";
+}
+
+export type GlossaryDeleteTermAssignmentFromEntitiesParameters = GlossaryDeleteTermAssignmentFromEntitiesMediaTypesParam &
+  GlossaryDeleteTermAssignmentFromEntitiesBodyParam &
+  RequestParameters;
 
 export interface GlossaryListRelatedTermsQueryParamProperties {
   /** The page size - by default there is no paging. */
@@ -483,7 +616,14 @@ export interface GlossaryUpdateGlossaryBodyParam {
   body: AtlasGlossary;
 }
 
-export type GlossaryUpdateGlossaryParameters = GlossaryUpdateGlossaryBodyParam & RequestParameters;
+export interface GlossaryUpdateGlossaryMediaTypesParam {
+  /** Request content type */
+  contentType?: "application/json";
+}
+
+export type GlossaryUpdateGlossaryParameters = GlossaryUpdateGlossaryMediaTypesParam &
+  GlossaryUpdateGlossaryBodyParam &
+  RequestParameters;
 export type GlossaryDeleteGlossaryParameters = RequestParameters;
 
 export interface GlossaryListGlossaryCategoriesQueryParamProperties {
@@ -515,8 +655,8 @@ export interface GlossaryListGlossaryCategoriesHeadersQueryParam {
   queryParameters?: GlossaryListGlossaryCategoriesHeadersQueryParamProperties;
 }
 
-export type GlossaryListGlossaryCategoriesHeadersParameters =
-  GlossaryListGlossaryCategoriesHeadersQueryParam & RequestParameters;
+export type GlossaryListGlossaryCategoriesHeadersParameters = GlossaryListGlossaryCategoriesHeadersQueryParam &
+  RequestParameters;
 
 export interface GlossaryGetDetailedGlossaryQueryParamProperties {
   /** Whether include term hierarchy */
@@ -544,7 +684,13 @@ export interface GlossaryPartialUpdateGlossaryQueryParam {
   queryParameters?: GlossaryPartialUpdateGlossaryQueryParamProperties;
 }
 
+export interface GlossaryPartialUpdateGlossaryMediaTypesParam {
+  /** Request content type */
+  contentType?: "application/json";
+}
+
 export type GlossaryPartialUpdateGlossaryParameters = GlossaryPartialUpdateGlossaryQueryParam &
+  GlossaryPartialUpdateGlossaryMediaTypesParam &
   GlossaryPartialUpdateGlossaryBodyParam &
   RequestParameters;
 
@@ -583,12 +729,16 @@ export type GlossaryListGlossaryTermHeadersParameters = GlossaryListGlossaryTerm
   RequestParameters;
 
 export interface GlossaryImportGlossaryTermsViaCsvBodyParam {
+  body: GlossaryImportGlossaryTermsViaCsvFormBody;
+}
+
+export interface GlossaryImportGlossaryTermsViaCsvFormBody {
   /**
    * The csv file to import glossary terms from.
    *
    * Value may contain any sequence of octets
    */
-  body: string;
+  file: string | Uint8Array;
 }
 
 export interface GlossaryImportGlossaryTermsViaCsvQueryParamProperties {
@@ -600,18 +750,27 @@ export interface GlossaryImportGlossaryTermsViaCsvQueryParam {
   queryParameters?: GlossaryImportGlossaryTermsViaCsvQueryParamProperties;
 }
 
-export type GlossaryImportGlossaryTermsViaCsvParameters =
-  GlossaryImportGlossaryTermsViaCsvQueryParam &
-    GlossaryImportGlossaryTermsViaCsvBodyParam &
-    RequestParameters;
+export interface GlossaryImportGlossaryTermsViaCsvMediaTypesParam {
+  /** Request content type */
+  contentType?: "multipart/form-data";
+}
+
+export type GlossaryImportGlossaryTermsViaCsvParameters = GlossaryImportGlossaryTermsViaCsvQueryParam &
+  GlossaryImportGlossaryTermsViaCsvMediaTypesParam &
+  GlossaryImportGlossaryTermsViaCsvBodyParam &
+  RequestParameters;
 
 export interface GlossaryImportGlossaryTermsViaCsvByGlossaryNameBodyParam {
+  body: GlossaryImportGlossaryTermsViaCsvByGlossaryNameFormBody;
+}
+
+export interface GlossaryImportGlossaryTermsViaCsvByGlossaryNameFormBody {
   /**
    * The csv file to import glossary terms from.
    *
    * Value may contain any sequence of octets
    */
-  body: string;
+  file: string | Uint8Array;
 }
 
 export interface GlossaryImportGlossaryTermsViaCsvByGlossaryNameQueryParamProperties {
@@ -623,10 +782,15 @@ export interface GlossaryImportGlossaryTermsViaCsvByGlossaryNameQueryParam {
   queryParameters?: GlossaryImportGlossaryTermsViaCsvByGlossaryNameQueryParamProperties;
 }
 
-export type GlossaryImportGlossaryTermsViaCsvByGlossaryNameParameters =
-  GlossaryImportGlossaryTermsViaCsvByGlossaryNameQueryParam &
-    GlossaryImportGlossaryTermsViaCsvByGlossaryNameBodyParam &
-    RequestParameters;
+export interface GlossaryImportGlossaryTermsViaCsvByGlossaryNameMediaTypesParam {
+  /** Request content type */
+  contentType?: "multipart/form-data";
+}
+
+export type GlossaryImportGlossaryTermsViaCsvByGlossaryNameParameters = GlossaryImportGlossaryTermsViaCsvByGlossaryNameQueryParam &
+  GlossaryImportGlossaryTermsViaCsvByGlossaryNameMediaTypesParam &
+  GlossaryImportGlossaryTermsViaCsvByGlossaryNameBodyParam &
+  RequestParameters;
 export type GlossaryGetImportCsvOperationStatusParameters = RequestParameters;
 
 export interface GlossaryExportGlossaryTermsAsCsvBodyParam {
@@ -643,10 +807,15 @@ export interface GlossaryExportGlossaryTermsAsCsvQueryParam {
   queryParameters?: GlossaryExportGlossaryTermsAsCsvQueryParamProperties;
 }
 
-export type GlossaryExportGlossaryTermsAsCsvParameters =
-  GlossaryExportGlossaryTermsAsCsvQueryParam &
-    GlossaryExportGlossaryTermsAsCsvBodyParam &
-    RequestParameters;
+export interface GlossaryExportGlossaryTermsAsCsvMediaTypesParam {
+  /** Request content type */
+  contentType?: "application/json";
+}
+
+export type GlossaryExportGlossaryTermsAsCsvParameters = GlossaryExportGlossaryTermsAsCsvQueryParam &
+  GlossaryExportGlossaryTermsAsCsvMediaTypesParam &
+  GlossaryExportGlossaryTermsAsCsvBodyParam &
+  RequestParameters;
 
 export interface GlossaryListTermsByGlossaryNameQueryParamProperties {
   /** The page size - by default there is no paging. */
@@ -669,28 +838,56 @@ export interface DiscoveryQueryBodyParam {
   body: SearchRequest;
 }
 
-export type DiscoveryQueryParameters = DiscoveryQueryBodyParam & RequestParameters;
+export interface DiscoveryQueryMediaTypesParam {
+  /** Request content type */
+  contentType?: "application/json";
+}
+
+export type DiscoveryQueryParameters = DiscoveryQueryMediaTypesParam &
+  DiscoveryQueryBodyParam &
+  RequestParameters;
 
 export interface DiscoverySuggestBodyParam {
   /** An object specifying the suggest criteria. */
   body: SuggestRequest;
 }
 
-export type DiscoverySuggestParameters = DiscoverySuggestBodyParam & RequestParameters;
+export interface DiscoverySuggestMediaTypesParam {
+  /** Request content type */
+  contentType?: "application/json";
+}
+
+export type DiscoverySuggestParameters = DiscoverySuggestMediaTypesParam &
+  DiscoverySuggestBodyParam &
+  RequestParameters;
 
 export interface DiscoveryBrowseBodyParam {
   /** An object specifying the browse criteria. */
   body: BrowseRequest;
 }
 
-export type DiscoveryBrowseParameters = DiscoveryBrowseBodyParam & RequestParameters;
+export interface DiscoveryBrowseMediaTypesParam {
+  /** Request content type */
+  contentType?: "application/json";
+}
+
+export type DiscoveryBrowseParameters = DiscoveryBrowseMediaTypesParam &
+  DiscoveryBrowseBodyParam &
+  RequestParameters;
 
 export interface DiscoveryAutoCompleteBodyParam {
   /** An object specifying the autocomplete criteria. */
   body: AutoCompleteRequest;
 }
 
-export type DiscoveryAutoCompleteParameters = DiscoveryAutoCompleteBodyParam & RequestParameters;
+export interface DiscoveryAutoCompleteMediaTypesParam {
+  /** Request content type */
+  contentType?: "application/json";
+}
+
+export type DiscoveryAutoCompleteParameters = DiscoveryAutoCompleteMediaTypesParam &
+  DiscoveryAutoCompleteBodyParam &
+  RequestParameters;
 
 export interface LineageGetLineageGraphQueryParamProperties {
   /** The number of hops for lineage. */
@@ -709,7 +906,8 @@ export interface LineageGetLineageGraphQueryParam {
   queryParameters: LineageGetLineageGraphQueryParamProperties;
 }
 
-export type LineageGetLineageGraphParameters = LineageGetLineageGraphQueryParam & RequestParameters;
+export type LineageGetLineageGraphParameters = LineageGetLineageGraphQueryParam &
+  RequestParameters;
 
 export interface LineageNextPageLineageQueryParamProperties {
   /** The direction of the lineage, which could be INPUT, OUTPUT or BOTH. */
@@ -726,21 +924,36 @@ export interface LineageNextPageLineageQueryParam {
   queryParameters: LineageNextPageLineageQueryParamProperties;
 }
 
-export type LineageNextPageLineageParameters = LineageNextPageLineageQueryParam & RequestParameters;
+export type LineageNextPageLineageParameters = LineageNextPageLineageQueryParam &
+  RequestParameters;
 
 export interface RelationshipCreateBodyParam {
   /** The AtlasRelationship object containing the information for the relationship to be created. */
   body: AtlasRelationship;
 }
 
-export type RelationshipCreateParameters = RelationshipCreateBodyParam & RequestParameters;
+export interface RelationshipCreateMediaTypesParam {
+  /** Request content type */
+  contentType?: "application/json";
+}
+
+export type RelationshipCreateParameters = RelationshipCreateMediaTypesParam &
+  RelationshipCreateBodyParam &
+  RequestParameters;
 
 export interface RelationshipUpdateBodyParam {
   /** The AtlasRelationship object containing the information for the relationship to be created. */
   body: AtlasRelationship;
 }
 
-export type RelationshipUpdateParameters = RelationshipUpdateBodyParam & RequestParameters;
+export interface RelationshipUpdateMediaTypesParam {
+  /** Request content type */
+  contentType?: "application/json";
+}
+
+export type RelationshipUpdateParameters = RelationshipUpdateMediaTypesParam &
+  RelationshipUpdateBodyParam &
+  RequestParameters;
 
 export interface RelationshipGetQueryParamProperties {
   /** Limits whether includes extended information. */
@@ -751,7 +964,8 @@ export interface RelationshipGetQueryParam {
   queryParameters?: RelationshipGetQueryParamProperties;
 }
 
-export type RelationshipGetParameters = RelationshipGetQueryParam & RequestParameters;
+export type RelationshipGetParameters = RelationshipGetQueryParam &
+  RequestParameters;
 export type RelationshipDeleteParameters = RequestParameters;
 export type TypesGetClassificationDefByGuidParameters = RequestParameters;
 export type TypesGetClassificationDefByNameParameters = RequestParameters;
@@ -774,7 +988,13 @@ export interface TypesGetAllTypeDefinitionsQueryParamProperties {
    */
   includeTermTemplate?: boolean;
   /** Typedef name as search filter when get typedefs. */
-  type?: "enum" | "entity" | "classification" | "relationship" | "struct" | "term_template";
+  type?:
+    | "enum"
+    | "entity"
+    | "classification"
+    | "relationship"
+    | "struct"
+    | "term_template";
 }
 
 export interface TypesGetAllTypeDefinitionsQueryParam {
@@ -789,7 +1009,13 @@ export interface TypesCreateTypeDefinitionsBodyParam {
   body: AtlasTypesDef;
 }
 
-export type TypesCreateTypeDefinitionsParameters = TypesCreateTypeDefinitionsBodyParam &
+export interface TypesCreateTypeDefinitionsMediaTypesParam {
+  /** Request content type */
+  contentType?: "application/json";
+}
+
+export type TypesCreateTypeDefinitionsParameters = TypesCreateTypeDefinitionsMediaTypesParam &
+  TypesCreateTypeDefinitionsBodyParam &
   RequestParameters;
 
 export interface TypesUpdateAtlasTypeDefinitionsBodyParam {
@@ -797,7 +1023,13 @@ export interface TypesUpdateAtlasTypeDefinitionsBodyParam {
   body: AtlasTypesDef;
 }
 
-export type TypesUpdateAtlasTypeDefinitionsParameters = TypesUpdateAtlasTypeDefinitionsBodyParam &
+export interface TypesUpdateAtlasTypeDefinitionsMediaTypesParam {
+  /** Request content type */
+  contentType?: "application/json";
+}
+
+export type TypesUpdateAtlasTypeDefinitionsParameters = TypesUpdateAtlasTypeDefinitionsMediaTypesParam &
+  TypesUpdateAtlasTypeDefinitionsBodyParam &
   RequestParameters;
 
 export interface TypesDeleteTypeDefinitionsBodyParam {
@@ -805,7 +1037,13 @@ export interface TypesDeleteTypeDefinitionsBodyParam {
   body: AtlasTypesDef;
 }
 
-export type TypesDeleteTypeDefinitionsParameters = TypesDeleteTypeDefinitionsBodyParam &
+export interface TypesDeleteTypeDefinitionsMediaTypesParam {
+  /** Request content type */
+  contentType?: "application/json";
+}
+
+export type TypesDeleteTypeDefinitionsParameters = TypesDeleteTypeDefinitionsMediaTypesParam &
+  TypesDeleteTypeDefinitionsBodyParam &
   RequestParameters;
 
 export interface TypesListTypeDefinitionHeadersQueryParamProperties {
@@ -815,7 +1053,13 @@ export interface TypesListTypeDefinitionHeadersQueryParamProperties {
    */
   includeTermTemplate?: boolean;
   /** Typedef name as search filter when get typedefs. */
-  type?: "enum" | "entity" | "classification" | "relationship" | "struct" | "term_template";
+  type?:
+    | "enum"
+    | "entity"
+    | "classification"
+    | "relationship"
+    | "struct"
+    | "term_template";
 }
 
 export interface TypesListTypeDefinitionHeadersQueryParam {
@@ -832,7 +1076,13 @@ export interface CollectionCreateOrUpdateBodyParam {
   body: AtlasEntityWithExtInfo;
 }
 
-export type CollectionCreateOrUpdateParameters = CollectionCreateOrUpdateBodyParam &
+export interface CollectionCreateOrUpdateMediaTypesParam {
+  /** Request content type */
+  contentType?: "application/json";
+}
+
+export type CollectionCreateOrUpdateParameters = CollectionCreateOrUpdateMediaTypesParam &
+  CollectionCreateOrUpdateBodyParam &
   RequestParameters;
 
 export interface CollectionCreateOrUpdateBulkBodyParam {
@@ -840,7 +1090,13 @@ export interface CollectionCreateOrUpdateBulkBodyParam {
   body: AtlasEntitiesWithExtInfo;
 }
 
-export type CollectionCreateOrUpdateBulkParameters = CollectionCreateOrUpdateBulkBodyParam &
+export interface CollectionCreateOrUpdateBulkMediaTypesParam {
+  /** Request content type */
+  contentType?: "application/json";
+}
+
+export type CollectionCreateOrUpdateBulkParameters = CollectionCreateOrUpdateBulkMediaTypesParam &
+  CollectionCreateOrUpdateBulkBodyParam &
   RequestParameters;
 
 export interface CollectionMoveEntitiesToCollectionBodyParam {
@@ -848,5 +1104,11 @@ export interface CollectionMoveEntitiesToCollectionBodyParam {
   body: MoveEntitiesRequest;
 }
 
-export type CollectionMoveEntitiesToCollectionParameters =
-  CollectionMoveEntitiesToCollectionBodyParam & RequestParameters;
+export interface CollectionMoveEntitiesToCollectionMediaTypesParam {
+  /** Request content type */
+  contentType?: "application/json";
+}
+
+export type CollectionMoveEntitiesToCollectionParameters = CollectionMoveEntitiesToCollectionMediaTypesParam &
+  CollectionMoveEntitiesToCollectionBodyParam &
+  RequestParameters;
